LPS-179353 Extract ScreenNavigationEntry methods from CPInstanceSubscriptionInfoScreenNavigationCategory to CPInstanceSubscriptionInfoScreenNavigationEntry, extends CPInstanceSubscriptionInfoScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
