+++ modules/apps/commerce/commerce-subscription-web/src/main/java/com/liferay/commerce/subscription/web/internal/frontend/taglib/servlet/taglib/CPInstanceSubscriptionInfoScreenNavigationCategory.java
@@ -14,36 +14,13 @@
 
 package com.liferay.commerce.subscription.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.product.model.CPInstance;
-import com.liferay.commerce.product.model.CommerceCatalog;
-import com.liferay.commerce.product.portlet.action.ActionHelper;
 import com.liferay.commerce.product.servlet.taglib.ui.constants.CPInstanceScreenNavigationConstants;
-import com.liferay.commerce.product.util.CPSubscriptionTypeJSPContributorRegistry;
-import com.liferay.commerce.product.util.CPSubscriptionTypeRegistry;
-import com.liferay.commerce.subscription.web.internal.display.context.CPInstanceSubscriptionInfoDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -53,23 +30,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CPInstanceSubscriptionInfoScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<CPInstance> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CPInstanceScreenNavigationConstants.
-			CATEGORY_KEY_SUBSCRIPTION_OVERRIDE;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CPInstanceScreenNavigationConstants.
 			CATEGORY_KEY_SUBSCRIPTION_OVERRIDE;
 	}
@@ -79,7 +47,7 @@ public class CPInstanceSubscriptionInfoScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, "subscriptions");
+		return language.get(resourceBundle, "subscriptions");
 	}
 
 	@Override
@@ -88,82 +56,7 @@ public class CPInstanceSubscriptionInfoScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_CP_INSTANCE_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CPInstance cpInstance) {
-		if (cpInstance == null) {
-			return false;
-		}
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		try {
-			return _commerceCatalogModelResourcePermission.contains(
-				permissionChecker, cpInstance.getCommerceCatalog(),
-				ActionKeys.VIEW);
-		}
-		catch (PortalException portalException) {
-			_log.error(portalException);
-		}
-
-		return false;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		try {
-			CPInstanceSubscriptionInfoDisplayContext
-				cpInstanceSubscriptionInfoDisplayContext =
-					new CPInstanceSubscriptionInfoDisplayContext(
-						_actionHelper, httpServletRequest,
-						_cpSubscriptionTypeJSPContributorRegistry,
-						_cpSubscriptionTypeRegistry);
-
-			httpServletRequest.setAttribute(
-				WebKeys.PORTLET_DISPLAY_CONTEXT,
-				cpInstanceSubscriptionInfoDisplayContext);
-		}
-		catch (Exception exception) {
-			_log.error(exception);
-		}
-
-		_jspRenderer.renderJSP(
-			_setServletContext, httpServletRequest, httpServletResponse,
-			"/definition/edit_instance_subscription_info.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CPInstanceSubscriptionInfoScreenNavigationCategory.class);
-
 	@Reference
-	private ActionHelper _actionHelper;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.product.model.CommerceCatalog)"
-	)
-	private ModelResourcePermission<CommerceCatalog>
-		_commerceCatalogModelResourcePermission;
-
-	@Reference
-	private CPSubscriptionTypeJSPContributorRegistry
-		_cpSubscriptionTypeJSPContributorRegistry;
-
-	@Reference
-	private CPSubscriptionTypeRegistry _cpSubscriptionTypeRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.subscription.web)"
-	)
-	private ServletContext _setServletContext;
+	protected Language language;
 
 }
