+++ modules/apps/commerce/commerce-subscription-web/src/main/java/com/liferay/commerce/subscription/web/internal/frontend/taglib/servlet/taglib/CPInstanceSubscriptionInfoScreenNavigationEntry.java
@@ -1,0 +1,135 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.subscription.web.internal.frontend.taglib.servlet.taglib;
+
+import com.liferay.commerce.product.model.CPInstance;
+import com.liferay.commerce.product.model.CommerceCatalog;
+import com.liferay.commerce.product.portlet.action.ActionHelper;
+import com.liferay.commerce.product.util.CPSubscriptionTypeJSPContributorRegistry;
+import com.liferay.commerce.product.util.CPSubscriptionTypeRegistry;
+import com.liferay.commerce.subscription.web.internal.display.context.CPInstanceSubscriptionInfoDisplayContext;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
+import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
+import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
+import com.liferay.portal.kernel.util.WebKeys;
+
+import java.io.IOException;
+
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(
+	property = "screen.navigation.entry.order:Integer=10",
+	service = ScreenNavigationEntry.class
+)
+public class CPInstanceSubscriptionInfoScreenNavigationEntry
+	extends CPInstanceSubscriptionInfoScreenNavigationCategory
+	implements ScreenNavigationEntry<CPInstance> {
+
+	@Override
+	public String getEntryKey() {
+		return getCategoryKey();
+	}
+
+	@Override
+	public boolean isVisible(User user, CPInstance cpInstance) {
+		if (cpInstance == null) {
+			return false;
+		}
+
+		PermissionChecker permissionChecker =
+			PermissionThreadLocal.getPermissionChecker();
+
+		try {
+			return _commerceCatalogModelResourcePermission.contains(
+				permissionChecker, cpInstance.getCommerceCatalog(),
+				ActionKeys.VIEW);
+		}
+		catch (PortalException portalException) {
+			_log.error(portalException);
+		}
+
+		return false;
+	}
+
+	@Override
+	public void render(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws IOException {
+
+		try {
+			CPInstanceSubscriptionInfoDisplayContext
+				cpInstanceSubscriptionInfoDisplayContext =
+					new CPInstanceSubscriptionInfoDisplayContext(
+						_actionHelper, httpServletRequest,
+						_cpSubscriptionTypeJSPContributorRegistry,
+						_cpSubscriptionTypeRegistry);
+
+			httpServletRequest.setAttribute(
+				WebKeys.PORTLET_DISPLAY_CONTEXT,
+				cpInstanceSubscriptionInfoDisplayContext);
+		}
+		catch (Exception exception) {
+			_log.error(exception);
+		}
+
+		_jspRenderer.renderJSP(
+			_setServletContext, httpServletRequest, httpServletResponse,
+			"/definition/edit_instance_subscription_info.jsp");
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		CPInstanceSubscriptionInfoScreenNavigationEntry.class);
+
+	@Reference
+	private ActionHelper _actionHelper;
+
+	@Reference(
+		target = "(model.class.name=com.liferay.commerce.product.model.CommerceCatalog)"
+	)
+	private ModelResourcePermission<CommerceCatalog>
+		_commerceCatalogModelResourcePermission;
+
+	@Reference
+	private CPSubscriptionTypeJSPContributorRegistry
+		_cpSubscriptionTypeJSPContributorRegistry;
+
+	@Reference
+	private CPSubscriptionTypeRegistry _cpSubscriptionTypeRegistry;
+
+	@Reference
+	private JSPRenderer _jspRenderer;
+
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.commerce.subscription.web)"
+	)
+	private ServletContext _setServletContext;
+
+}
